backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-autoslug==1.9.8
django-cors-headers==4.6.0
django-filter==24.3z`
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==9.1.1
PyJWT==2.10.1
sqlparse==0.5.3
Unidecode==1.3.4
